Refactor: deduplicate test helpers and SHA accelerator code
- Add write_data_to_datain() helper in SHA accelerator to eliminate
  duplicated dword-writing loop in digest_384_slice, digest_384_slices,
  digest_512_slice, and digest_512_slices functions

- Consolidate test helpers in caliptra-test crate:
  - Export DEFAULT_MCU_FW from caliptra_auth_man_gen
  - Add shared default_soc_manifest_bytes() function
  - Add shared test_upload_firmware() generic function

- Update all test modules to use shared helpers:
  - test/tests/caliptra_integration_tests/{smoke_test,warm_reset}.rs
  - fmc/tests/fmc_integration_tests/test_rtalias.rs
  - runtime/tests/runtime_integration_tests/{common,test_info,test_warm_reset}.rs
  - rom/dev/tests/rom_integration_tests/helpers.rs
